Define #reveal, #reveal_adjacents Correct #render
Board #reveal will reveal the value of a Tile if it is not flagged. If the Tile
is a bomb or flagged, nothing happens. If the Tile has no bombs adjacent to it
it will run #reveal_adjacents on all adjacent Tiles. Otherwise it displays the
number of adjacent bombs. Yay, recursion!

Board#render returned the grid. That could be used to cheat. tsk, tsk.